Fix another printf format warning with qsizetype
As done in 939f5112a1033299ec1a178fbd4232717a98e735 elsewhere, use %zd
and cast explicitly to size_t.

+++ src/tools/androiddeployqt/main.cpp
@@ -1195,7 +1195,7 @@ bool copyAndroidExtraResources(Options *options)
         return true;
 
     if (options->verbose)
-        fprintf(stdout, "Copying %d external resources to package.\n", options->extraPlugins.size());
+        fprintf(stdout, "Copying %zd external resources to package.\n", size_t(options->extraPlugins.size()));
 
     for (const QString &extraResource : options->extraPlugins) {
         QFileInfo extraResourceInfo(extraResource);
